test: refine shared UI coverage review fixes (#897)
